refactor(hooks): extract safePath to shared module
The identical 5-line safePath function was duplicated in json-helper.cjs
and sidecar-ops.cjs. Extracted to scripts/hooks/lib/safe-path.cjs as a
single source of truth, with both consumers importing from it.
